+++ Backend/.env
@@ -1,6 +1,6 @@
 DB_NAME=innlabdb
 DB_USER=postgres
-DB_PASSWORD=postgres
+DB_PASSWORD=AxelitoSQL
 DB_HOST=localhost
 DB_DIALECT=postgres
 DB_PORT=5432
