+++ Frontend/src/components/Buttons/NextButton/next-button.css
